separer i flere filer for enklere oversikt
